update long lat

+++ src/main/java/com/project2/wanderfun/domain/model/Album.java
@@ -8,8 +8,8 @@ public class Album {
     private String name;
     private String description;
     private Long placeId;
-    private BigDecimal placeLongitude;
-    private BigDecimal placeLatitude;
+    private double placeLongitude;
+    private double placeLatitude;
     private String placeName;
     private String placeCoverImageUrl;
     private List<AlbumImage> albumImages;
@@ -50,19 +50,19 @@ public class Album {
         this.placeId = placeId;
     }
 
-    public BigDecimal getPlaceLongitude() {
+    public double getPlaceLongitude() {
         return placeLongitude;
     }
 
-    public void setPlaceLongitude(BigDecimal placeLongitude) {
+    public void setPlaceLongitude(double placeLongitude) {
         this.placeLongitude = placeLongitude;
     }
 
-    public BigDecimal getPlaceLatitude() {
+    public double getPlaceLatitude() {
         return placeLatitude;
     }
 
-    public void setPlaceLatitude(BigDecimal placeLatitude) {
+    public void setPlaceLatitude(double placeLatitude) {
         this.placeLatitude = placeLatitude;
     }
